using moments
Here I changed the slightly changed the code, instead of using
HoughCircle function now I am using Moments to calculate the offset
amount from the center of image followed by an if condition. Which is if
the number of white pixels is greater than some value say 4000, then
give the command to the bot. Here 4000 is the number of white pixels,
this is just avoid external noise. It can be any number but choose the
such that noise is not affecting the detection.